+++ WPF_PAR.Core/Models/SubLineaPerformanceModel.cs
@@ -8,14 +8,14 @@ namespace WPF_PAR.Core.Models
     {
         public string Etiqueta { get; set; }
         public decimal Valor { get; set; }
-        public double Litros { get; set; }
+        public decimal Litros { get; set; }
         public bool EsFuturo { get; set; }
     }
     public class SubLineaPerformanceModel
     {
         public string Nombre { get; set; }
         public decimal VentaTotal { get; set; }
-        public double LitrosTotales { get; set; }
+        public decimal LitrosTotales { get; set; }
         public List<PeriodoBloque> Bloques { get; set; }
     }
 }
